refactor: Move finish exam functionality to socket
- integrate `Score` model and `finishExam` controller function in `socket.exam.js`.
- convert HTTP protocol features:
function parameters from `req` and `res` to `context` and `data`, handling invalid data and errors by throwing new `Errors` instead of utilizing `res` object, emitting `error` event in `catch` code block in `finishExam` in `socket.exam.js`.
- remove integration of `finishExam` and its respective API route endpoint in `examRoutes.js`.
