+++ backend/controller/socket/helpers/socket.exam.js
@@ -1,6 +1,7 @@
 import Exam from "../../../models/quizzes/examModel.js";
 import Quiz from "../../../models/quizzes/quizModel.js";
 import Classroom from "../../../models/classrooms/classroomModel.js";
+import Score from "../../../models/quizzes/scoreModel.js";
 
 export const createExam = async (context, data) => {
   try {
@@ -105,5 +106,96 @@ export const updateExam = async (context, data) => {
     context.emitEvent("sender", "error", {
       message: `Error updating exam ${data?.receiverId}:  ${error.message}`,
     });
+  }
+};
+
+export const finishExam = async (data) => {
+  try {
+    const { senderId, quizId } = data;
+
+    const examFound = await Exam.findOne({ studentId: senderId });
+
+    if (!examFound) {
+      throw new Error("Exam not found");
+    }
+
+    const examQuestions = examFound?.shuffledQuestions;
+
+    const examIsValid = examFound?.examFinish?.getTime() - Date.now() > 0;
+
+    if (!examIsValid) {
+      throw new Error("Exam not found");
+    }
+
+    await Exam.findByIdAndUpdate(
+      examFound?._id,
+      { isInProgress: false },
+      { new: true }
+    );
+
+    const quizFound = await Quiz.findOne({ _id: quizId });
+
+    if (!quizFound) {
+      throw new Error("Cannot evaluate exam, no matching quiz found");
+    }
+
+    let currentScore = 0;
+
+    let userChoices = [];
+
+    for (let i = 0; i < quizFound.questions.length; i++) {
+      const q = quizFound.questions[i];
+
+      if (q.answer === examFound.answers[i]) {
+        currentScore += 1;
+      }
+
+      userChoices[i] = {
+        userAnswer: examFound?.answers?.[i],
+        correctAnswer: q.answer,
+      };
+    }
+
+    const scoreFound = await Score.findOne({
+      user: senderId,
+      quiz: quizFound?._id,
+    });
+
+    let scorePayload;
+
+    if (scoreFound) {
+      scorePayload = await Score.findByIdAndUpdate(
+        scoreFound?._id,
+        {
+          $set: {
+            "examFeedback.userChoices": userChoices,
+            "examFeedback.randomizedQuestions": examQuestions,
+            highScore: Math.max(scoreFound.highScore, currentScore),
+            latestScore: currentScore,
+          },
+        },
+        { new: true }
+      );
+    } else {
+      scorePayload = new Score({
+        user: senderId,
+        quiz: quizFound?._id,
+        examFeedback: {
+          userChoices,
+          randomizedQuestions: examQuestions,
+        },
+        highScore: currentScore,
+        latestScore: currentScore,
+      }).save();
+    }
+
+    const examPayload = { examId: examFound?._id, scorePayload };
+
+    await Exam.findByIdAndDelete({ _id: examFound?._id });
+
+    return { success: true, examPayload };
+  } catch (error) {
+    console.error(`Error fetching exam feedback: ${error.message}`);
+    return { success: false, message: `Error finishing exam ${error.message}` };
   }
 };
